Use base image's kernel in image prep phase 2
Previously (in the pre-release Buster and a few first release
kernels) the support for RTC clock was disabled in cloud images,
and we had to replace the base image kernel with a matching non-
cloud, regular image. Now they added the support, so there is no
longer a reason for that. We always compile NVIDIA drivers for
the running kernel, and do not allow kernel upgrade during the
apt upgrade when preparing the image. Google always supplies fresh
and very recent software in the base image. NVIDIA build is now
significantly more robust.

Close: #12