Fix 'bm-os-image rollout' broken by DM backend change
The command broke suddenly with an ultimately unhelphul error
message. The root cause was the change in Compute API that no
longer accepts a PUT request removing a boot disk from an
instance, if the same request also contains shieldedInstanceConfig
object, complaining that it is allowed only for UEFI-bootable
instances--as if the instance without a boot disk were bootable
any other way; the added check makes little sense, IMO.

DM is still in beta, but the Compute API v1, which it calls and
which returns the error, is GA, and is supposed to be stable.
This is nasty.

To apply this fix, the users will need to rebuild the cluster
deployment record with 'bm-deploy fix', since the template is
baked into the deployent manifest, and we do not do updates of
the template. Scattering and gathering parts fixes it just as
fine, so handling this as a separate case is not really feasible.