Retry jobs exited with status 127 up to 4 times (#17)
This is a workaround for the problem in issue #17.

Jobs randomly, and very rarely, fail with exit code 127 and
a message from ld.so "PROGNAME: error while loading shared
libraries: libLIBNAME.so: cannot open shared object file:
No such file or directory" when loading from the CNS disk
(the /opt mount).

Issue: #17