Add ensN to pattern of Ether drivers in udevd rule
The 'ens5' was not recognized by 71-ether_txquelen.rules file which
extends tx_queue_len to 2500 on the matching interface.

Although we normally disable "persistent" interface names in the
full Burrmill setup in favor of good old ethN, because they aren't
persistent in GCE and change with an addition or removal of GPU
accelerators, the rule file may be used for a setup independent of
Burrmill.

The authors in fact use it for unrelated high-throughput setups. We
do not intend to cover all possible patterns; anything but eth[0-9]*
is supported ad-hoc entirely for our own convenience.